# openapi.yaml
openapi: 3.0.3
info:
  title: Bambuzle API
  description: |
    REST API for the Bambuzle 3D printer monitoring dashboard.

    Bambuzle connects to BambuLab printers via MQTT, collects telemetry (temperatures, layer progress, fan speeds),
    tracks print jobs, detects anomalies, and provides real-time state via WebSocket.

    ## Authentication
    Two separate layers:
    1. **Dashboard admin token** (who may use this API). Every `/api/*` route and the WebSocket require it, except
       `/api/session`, `/api/spec` and `/api/docs` (and the `/healthz` / `/readyz` probes, which live outside `/api`). Send `Authorization: Bearer <token>`, or exchange the token for an
       HttpOnly session cookie via `POST /api/session`. `BAMBUZLE_PUBLIC_READ=true` opens GETs and the WebSocket;
       `BAMBUZLE_AUTH=off` disables the check entirely.
    2. **BambuLab Cloud login** (the server's own connection to your printers), via `/api/auth/login`.

    ## Real-time Data
    For live updates, connect to the WebSocket endpoint at `ws://<host>:<port>/ws`.
    Messages are JSON with `{ type, data }` format. Types include `state`, `event`, and `auth`.
  version: 0.7.0
  contact:
    name: Bambuzle
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development
security:
  - bearerAuth: []
  - sessionCookie: []
...
paths:
...
  /api/system:
    get:
      tags: [System]
      summary: Version, uptime, database size/row counts, backup schedule and last result
      responses:
        '200':
          description: System info
          content:
            application/json:
              schema:
                type: object
                properties:
                  version: { type: string }
                  node: { type: string }
                  uptimeSec: { type: integer }
                  dataDir: { type: string }
                  db:
                    type: object
                    properties:
                      path: { type: string }
                      sizeBytes: { type: integer }
                      walBytes: { type: integer }
                      rowCounts:
                        type: object
                        properties:
                          samples: { type: integer }
                          events: { type: integer }
                          print_jobs: { type: integer }
                  backup:
                    type: object
                    properties:
                      schedule:
                        type: object
                        properties:
                          enabled: { type: boolean }
                          cron: { type: string }
                          dir: { type: string }
                          keep: { type: integer }
                          nextRun: { type: string, format: date-time, nullable: true }
                      last:
                        nullable: true
                        $ref: '#/components/schemas/BackupResult'
        '401':
          description: Admin token required
...
  /api/alerts:
    get:
      tags: [Alerts]
      summary: List all alert rules
      responses:
        '200':
          description: Array of alert rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AlertRule'
    post:
      tags: [Alerts]
      summary: Create an alert rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertRuleInput'
      responses:
        '201':
          description: Created alert rule
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertRule'
        '400':
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/alerts/{alertId}:
    get:
      tags: [Alerts]
      summary: Get a single alert rule
      parameters:
        - name: alertId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Alert rule
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertRule'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags: [Alerts]
      summary: Update an alert rule
      parameters:
        - name: alertId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertRuleInput'
      responses:
        '200':
          description: Updated alert rule
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertRule'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [Alerts]
      summary: Delete an alert rule
      parameters:
        - name: alertId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ─── Debug ───
...
  /api/audit:
    get:
      tags: [Audit]
      summary: Operator audit trail, newest first
      description: |
        BAM-41. Append-only record of security-relevant and state-changing operator actions: dashboard sign-in /
        sign-out / rate-limit hits, admin-guard denials on mutating requests (throttled to one row per source IP +
        action per minute; the next row carries `detail.suppressed`), BambuLab Cloud login / verify / logout,
        connection settings, LAN printer add / remove, connection tests, printer commands, alert rule and
        maintenance task changes, manual backups.

        Private even with `BAMBUZLE_PUBLIC_READ=true`. `detail` never contains secrets (access codes, passwords,
        verification codes, tokens, notifier config) — only field names and flags such as `accessCodeChanged`.
        Rows older than `audit.retentionDays` (default 365, env `BAMBUZLE_AUDIT_RETENTION_DAYS`, 0 = forever) are
        pruned by the daily cleanup.
      parameters:
        - $ref: '#/components/parameters/AuditFrom'
        - $ref: '#/components/parameters/AuditTo'
        - $ref: '#/components/parameters/AuditAction'
        - $ref: '#/components/parameters/AuditResult'
        - name: limit
          in: query
          schema:
            type: integer
            default: 500
            minimum: 1
            maximum: 2000
      responses:
        '200':
          description: Matching entries (at most `limit`)
          content:
            application/json:
              schema:
                type: object
                properties:
                  window:
                    type: object
                    properties:
                      from: { type: string, format: date-time }
                      to: { type: string, format: date-time }
                  filters:
                    type: object
                    properties:
                      action: { type: string, nullable: true }
                      result: { type: string, nullable: true }
                  limit: { type: integer }
                  truncated:
                    type: boolean
                    description: True when `limit` entries were returned (there may be more)
                  entries:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditEntry'
        '400':
          description: Bad from/to/action/result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/MaintUnauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Dashboard admin token (BAMBUZLE_ADMIN_TOKEN or <data dir>/admin-token)
    sessionCookie:
      type: apiKey
      in: cookie
      name: bambuzle_session
  parameters:
...
    AuditFrom:
      name: from
      in: query
      description: Window start (ISO 8601 date or date-time; offset-less values are UTC). Default 30 days before `to`.
      schema:
        type: string
        example: '2026-09-01'
    AuditTo:
      name: to
      in: query
      description: Window end (ISO 8601; a bare date means the end of that UTC day). Default now.
      schema:
        type: string
    AuditAction:
      name: action
      in: query
      description: Exact action (`printer.command`) or category prefix (`printer` matches `printer.*`)
      schema:
        type: string
        example: printer.command
    AuditResult:
      name: result
      in: query
      schema:
        type: string
        enum: [ok, denied, rejected, error]
  responses:
...
    MaintUnauthorized:
      description: Missing or invalid admin token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    AuditEntry:
      type: object
      properties:
        id: { type: integer }
        ts: { type: string, format: date-time, description: ISO 8601 UTC with milliseconds }
        actor:
          type: string
          description: Credential that authorised the request — `admin-token` (Bearer), `session` (cookie) or `anonymous`. Per-user identities arrive with BAM-16.
        source_ip: { type: string, nullable: true, description: Client IP (honours BAMBUZLE_TRUST_PROXY) }
        user_agent: { type: string, nullable: true, maxLength: 200 }
        action:
          type: string
          description: |
            session.login, session.logout, session.rate_limited, access.denied, access.cross_origin,
            cloud.login, cloud.verify, cloud.logout, printer.connection.update, printer.connection.test,
            printer.add, printer.remove, printer.command, alert.create, alert.update, alert.delete,
            maintenance.task.create, maintenance.task.update, maintenance.task.delete, maintenance.task.done,
            maintenance.templates, system.backup
        target: { type: string, nullable: true, example: 'printer:01S00EXAMPLE01' }
        result:
          type: string
          enum: [ok, denied, rejected, error]
          description: ok = done; denied = authentication/authorisation refused; rejected = validation, state gate or printer refusal; error = failure (incl. unconfirmed commands)
        detail:
          type: object
          nullable: true
          description: Small secret-free JSON, e.g. `{"accessCodeChanged":true}` or `{"command":"pause","outcome":"confirmed"}`
        request_id: { type: string, nullable: true, description: Also sent as the `X-Request-Id` response header }
    Error:
      type: object
      properties:
        error:
          type: string
...
    BackupResult:
      type: object
      properties:
        ok: { type: boolean, nullable: true }
        path: { type: string, nullable: true }
        time: { type: string, format: date-time }
        size: { type: integer, nullable: true }
        sha256: { type: string }
        durationMs: { type: integer }
        pruned: { type: integer }
        error: { type: string }
        source: { type: string, enum: [disk], description: Present when seeded from the newest file on disk at startup }
...
    AlertRule:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        enabled:
          type: integer
          description: "1 = enabled, 0 = disabled"
        device_id:
          type: string
          nullable: true
          description: Null = applies to all printers
        condition_type:
          type: string
        condition_config:
          type: object
        notify_via:
          type: string
        notify_config:
          type: object
        cooldown_sec:
          type: integer
        last_fired_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    AlertRuleInput:
      type: object
      required: [name, conditionType]
      properties:
        name:
          type: string
        deviceId:
          type: string
          nullable: true
        conditionType:
          type: string
          enum: [state_change, hms_error, temp_anomaly, temp_threshold, progress_stall, print_error, ams_humidity]
        conditionConfig:
          type: object
        notifyVia:
          type: string
          enum: [console, webhook, ntfy, pushover, telegram]
          default: console
        notifyConfig:
          type: object
          description: |
            Per channel — webhook `{url, format: generic|slack|discord}`; ntfy `{server?, topic, token?}`;
            pushover `{appToken, userKey}`; telegram `{botToken, chatId}`.
        cooldownSec:
          type: integer
          default: 300

    # ─── Maintenance (BAM-39) ───
...